rewrite CLI mechanism to enable modular subcommands
- maintain backward compatibility
- add placeholder commands for `init`, `fmt`, `test`
