+++ src/commands/index.ts
@@ -1,0 +1,4 @@
+export * from "./build";
+export * from "./fmt";
+export * from "./test";
+export * from "./init";
