+++ src/commands/test.ts
@@ -1,0 +1,9 @@
+import * as yargs from "yargs";
+
+export const TestCmd: yargs.CommandModule = {
+  command: "test",
+  describe: "Execute as-pect tests of a local package",
+  handler: (args) => {
+    console.log(args);
+  },
+};
